Update samples to work with 64 bit
git-svn-id: https://svn.code.sf.net/p/directshownet/code/trunk@1036 b2a09c84-0b36-484d-b36e-ffab392acda3

+++ directshowlib/Samples/Players/DxPlay/Capture.cs
@@ -131,7 +131,7 @@ namespace DxPlay
             const int E_ABORT = unchecked((int)0x80004004);
 
             int hr;
-            int p1, p2;
+            IntPtr p1, p2;
             EventCode ec;
 
             do
